requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
future==0.18.2
psycopg2==2.8.6
pytz==2021.1
serializers==0.2.4
sqlparse==0.4.1
